Move ErrorHandlingTestCase to a new file.
