+++ payment-circle/src/test/kotlin/org/stellar/anchor/paymentservice/circle/CirclePaymentServiceTest.kt
@@ -32,53 +32,6 @@ import reactor.netty.ByteBufMono
 import reactor.netty.http.client.HttpClientResponse
 import shadow.com.google.common.reflect.TypeToken
 
-private class ErrorHandlingTestCase {
-  val requestMono: Mono<*>
-  var mockResponses: List<MockResponse>? = null
-    private set
-  var mockResponsesMap: Map<String, MockResponse>? = null
-    private set
-
-  constructor(_requestMono: Mono<*>, _mockResponses: List<MockResponse>) {
-    this.requestMono = _requestMono
-    this.mockResponses = _mockResponses
-  }
-
-  constructor(_requestMono: Mono<*>, _mockResponsesMap: Map<String, MockResponse>) {
-    this.requestMono = _requestMono
-    this.mockResponsesMap = _mockResponsesMap
-  }
-
-  private fun getDispatcher(): Dispatcher? {
-    if (mockResponsesMap == null) {
-      return null
-    }
-
-    val dispatcher: Dispatcher =
-      object : Dispatcher() {
-        @Throws(InterruptedException::class)
-        override fun dispatch(request: RecordedRequest): MockResponse {
-
-          if (!mockResponsesMap!!.containsKey(request.path)) {
-            return MockResponse().setResponseCode(404)
-          }
-
-          return mockResponsesMap!![request.path]!!
-        }
-      }
-    return dispatcher
-  }
-
-  fun prepareMockWebServer(server: MockWebServer) {
-    val dispatcher = getDispatcher()
-    if (dispatcher != null) {
-      server.dispatcher = dispatcher
-    } else if (mockResponses != null) {
-      mockResponses!!.forEach { mockResponse -> server.enqueue(mockResponse) }
-    }
-  }
-}
-
 class CirclePaymentServiceTest {
   private lateinit var server: MockWebServer
   private lateinit var service: PaymentService
